Dom adapter files renaming
