+++ src/client/dom-processor/dom-processor.js
@@ -1,4 +1,4 @@
-import DomAdapter from './adapter-client';
+import DomAdapter from './client-dom-adapter';
 import DomProcessor from '../../processing/dom/index';
 
 export default new DomProcessor(new DomAdapter());
